Create Zoo.cpp
You are required to enter a word that consists of and that denotes the number of Zs and Os respectively. The input word is considered similar to the word zoo if 2*x=y.
Determine if the entered word is similar to the word zoo.
For example, words such as zzoooo and zzzoooooo are similar to the word zoo but not the words such as zzooo and zzzooooo.